Don't try to include adapter API if transactions aren't supported
Previously, requiring dm-transactions when an adapter
that doesn't support transactions was in use, failed
with an ugly error. This patch makes dm-transactions
ignore that situation silently, by simply not trying
to include any adapter API extensions it obviously
can't find if the adapter doesn't support transactions.

Of course, no transaction functionality is available
to applications that use transactions with an adapter
that doesn't support that concept.